refactor: common components
