+++ ddd_modeling.md
@@ -203,17 +203,3 @@ findByUrl(url: string, userId: string): Promise<NewsEntity | null>
 本プロジェクトは**実用的な保守性**を**理論的なDDDの純粋性**よりも優先した設計となっており、コンテンツ管理システムのような比較的シンプルなドメインロジックには適している。しかし、より複雑なビジネスシナリオに対応するためには、リッチドメインモデリングの導入が望ましい。
 
 戦術的DDDパターン（リポジトリ、ドメインサービス、値オブジェクト）は堅実に実装されているが、戦略的DDDパターン（集約、ドメインイベント、複雑なビジネスロジック）の導入により、より表現力豊かで拡張性の高いドメインモデルを実現できる。
-
-
-│ │                                                                                                                                                                                                                                                           │ │
-│ │ 3. コンポーネント構成の最適化（オプション）                                                                                                                                                                                                               │ │
-│ │                                                                                                                                                                                                                                                           │ │
-│ │ /src/components/                                                                                                                                                                                                                                          │ │
-│ │ ├── features/    # 機能単位のコンポーネント構成│ │
-│ │ ├── domains/     # ドメイン固有コンポーネント│ │
-│ │ └── common/      # 共通UIコンポーネント│ │
-│ │                                                                                                                                                                                                                                                           │ │
-│ │ 4. 型定義の整理│ │
-│ ││ │
-│ │ - 共有型を/src/common/types/に集約│ │
-│ │ - ドメイン固有の型は各ドメインディレクトリに保持
