+++ src/components/common/card/link-card.tsx
@@ -1,7 +1,6 @@
 "use client";
 import ReactMarkdown from "react-markdown";
 import { ServerAction } from "@/common/types";
-import { validateAndNormalizeUrl } from "@/common/validate/validate-url";
 import { DeleteButtonWithModal } from "@/components/common/delete-button-with-modal";
 import { Badge } from "@/components/common/ui/badge";
 import {
@@ -11,6 +10,7 @@ import {
 	CardHeader,
 	CardTitle,
 } from "@/components/common/ui/card";
+import { validateAndNormalizeUrl } from "@/components/common/utils/validate-url";
 import { Link } from "@/i18n/routing";
 
 export type LinkCardData = {
