+++ src/components/common/search/searchable-card-layout.tsx
@@ -1,7 +1,6 @@
 "use client";
 import { useTranslations } from "next-intl";
 import type { ReactNode } from "react";
-import Loading from "@/components/common/loading";
 import { Input } from "@/components/common/ui/input";
 import { StatusCodeView } from "../status/status-code-view";
 
@@ -11,7 +10,6 @@ type SearchableCardLayoutProps<T> = {
 	handleSearchChange: (e: React.ChangeEvent<HTMLInputElement>) => Promise<void>;
 	renderCard: (item: T, index: number) => ReactNode;
 	gridClassName: string;
-	isLoading?: boolean;
 };
 
 export function SearchableCardLayout<T>({
@@ -20,11 +18,8 @@ export function SearchableCardLayout<T>({
 	handleSearchChange,
 	renderCard,
 	gridClassName,
-	isLoading = false,
 }: SearchableCardLayoutProps<T>) {
 	const t = useTranslations("label");
-
-	if (isLoading) return <Loading />;
 
 	return (
 		<div className="px-2 sm:px-4">
