+++ src/components/common/ui/dialog.tsx
@@ -4,7 +4,7 @@ import { Cross2Icon } from "@radix-ui/react-icons";
 import { Dialog as DialogPrimitive } from "radix-ui";
 import * as React from "react";
 
-import { cn } from "@/common/tailwind/utils";
+import { cn } from "@/components/common/utils/cn";
 
 const Dialog = DialogPrimitive.Root;
