+++ src/components/common/ui/tabs.tsx
@@ -3,7 +3,7 @@
 import { Tabs as TabsPrimitive } from "radix-ui";
 import * as React from "react";
 
-import { cn } from "@/common/tailwind/utils";
+import { cn } from "@/components/common/utils/cn";
 
 const Tabs = TabsPrimitive.Root;
